server/serverpl/requirements.txt
gitpython
django-markdown-deux
django
mock
gunicorn
dj-database-url
whitenoise
psycopg2
oauth2
django-bootstrap3
jsonfield
requests
jinja2
pytest
pytest-django
sympy
yattag
markdown
pytest-selenium
sphinx
recommonmark
